fix(activerecord): extract attribute-inspection.ts to break circular dependency
Move InspectionMask, inspectionFilter, and formatForInspect into a shared
attribute-inspection.ts module so core.ts and attribute-methods.ts no longer
import each other:
- core.ts: imports from attribute-inspection.ts, re-exports InspectionMask
- attribute-methods.ts: imports from attribute-inspection.ts, re-exports formatForInspect

Also:
- _updateRecord: add hasChangesToSave guard before writing timestamps,
  matching Rails should_record_timestamps? check
- base.ts: remove stale 'No silent-filter needed' comment (attributesForUpdate
  now explicitly filters readonly/counter-cache/virtual columns)
